requirements.txt
fastapi
uvicorn
python-dotenv
requests
confluent-kafka
kafka-python
